sdk/pyproject.toml
[project]
name = "clado-observe"
version = "0.1.0"
description = "Clado Observe"
readme = "README.md"
authors = [{ name = "Clado", email = "david@clado.ai" }]
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["clado", "observe", "sdk"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "python-dotenv>=1.0.1",
  "browser-use>=0.1.37",
  "anthropic>=0.38.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "flake8>=7.1.1",
]

[project.urls]
Homepage = "https://github.com/clado-ai/observability"
Issues = "https://github.com/clado-ai/observability/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clado_observe"]
include = ["/src/clado_observe", "/src/clado_observe/**/*.py"]

[tool.hatch.build]
directory = "dist"
artifacts = ["src/clado_observe/**/*.py"]
exclude = ["src/tests/*"]

[tool.ruff]
exclude = ["docs"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["clado_observe"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = ["docs", "build", "dist", ".venv", ".eggs"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = false
mypy_path = ["src"]
